Renames file_parsing->delimited_to_arrays
