+++ src/_arraykit.c
@@ -7,7 +7,7 @@
 
 # include "array_go.h"
 # include "block_index.h"
-# include "file_parsing.h"
+# include "delimited_to_arrays.h"
 # include "methods.h"
 # include "tri_map.h"
