+++ src/delimited_to_arrays.c
@@ -8,7 +8,7 @@
 # include "numpy/arrayobject.h"
 # include "numpy/halffloat.h"
 
-# include "file_parsing.h"
+# include "delimited_to_arrays.h"
 # include "utilities.h"
 
 #define AK_is_digit(c) (((unsigned)(c) - '0') < 10u)
